+++ src/themes/elegant/_theme.php
@@ -150,7 +150,8 @@ function getPageLinksHTML(at\nerdreich\wiki\WikiUI $wiki): string
 {
     $html = '';
     foreach ($wiki->getMenuItems() as $action => $label) {
-        $html .= '<a href="?' . $action . '">' . ___($label) . '</a><br>';
+        $html .= '<a class="nav-' . preg_replace('/.*=/', '', $action) . '" href="?' . $action . '">' .
+            ___($label) . '</a><br>';
     }
     return $html;
 }