+++ src/themes/elegant/js/main.js
@@ -13,7 +13,7 @@ function setStoreValue (key, value) {
 // -----------------------------------------------------------------------------
 
 function rememberAuthor () {
-  const form = document.querySelector('.elegant-edit form')
+  const form = document.querySelector('.page-edit form')
   if (form) {
     const author = getStoreValue('author')
     if (author) {